refactor code to utils and add import guards
