+++ tuning/acceleration/plugins/framework_plugin_autogptq.py
@@ -1,54 +1,17 @@
 
-def parse_configuration():
-    pass
-
 
 from transformers import TrainingArguments
 from peft import LoraConfig, prepare_model_for_kbit_training
 
 from types import MethodType
 
-from auto_gptq import AutoGPTQForCausalLM, BaseQuantizeConfig
-from auto_gptq.utils.peft_utils import get_gptq_peft_model
-from auto_gptq.utils.peft_utils import GPTQLoraModel
-from auto_gptq.nn_modules.qlinear.qlinear_tritonv2 import QuantLinear as triton_Qlinear
-
-from peft.tuners.lora.gptq import QuantLinear as LoraLinearGPTQ
 from peft.tuners.lora.model import LoraModel
 
-from typing import Tuple, List, Dict
+from typing import Tuple, Dict
 
 from .framework_plugin import AccelerationPlugin
 
 import torch
-
-def _replace_module(self, parent_module, child_name, new_module, old_module):
-
-    # replace the lora linear
-    setattr(parent_module, child_name, new_module)
-
-    # dispatch to correct device 
-    # FIXME: refactor
-    for name, module in new_module.named_modules():
-        if "lora_" in name:
-            device = (list(old_module.parameters()) + list(old_module.buffers()))[0].device
-            module.to(device)
-
-# def _dispatch_lora_linear_for_gptq(
-def _create_new_module_triton(
-    lora_config: LoraConfig,
-    adapter_name: str,
-    target: torch.nn.Module,
-    **kwargs,
-):
-    # if the base layer module matches a supported class, dispatch the lora linear
-    # to be installed
-    new_module = None
-    if isinstance(target, triton_Qlinear):
-        new_module = LoraLinearGPTQ(target, adapter_name, lora_config=lora_config, **kwargs)
-
-    # if module cannot be found, return None which results in a raise in the call-stack
-    return new_module
 
 class AutoGPTQAccelerationPlugin(AccelerationPlugin):
     
@@ -65,6 +28,9 @@ class AutoGPTQAccelerationPlugin(AccelerationPlugin):
         self._check_config_equal(key="peft.quantization.auto_gptq.from_quantized", value=True)
 
     def model_loader(self, model_name: str, **kwargs):
+
+        # guarded imports
+        from auto_gptq import AutoGPTQForCausalLM, BaseQuantizeConfig
 
         # Currently we allow only a quantized checkpoint to be loaded, we do not
         # implement the quantization process here. 
@@ -127,6 +93,11 @@ class AutoGPTQAccelerationPlugin(AccelerationPlugin):
         train_args: TrainingArguments,
         modifiable_args: Tuple[LoraConfig],
     ):
+        # guarded imports
+        from auto_gptq.utils.peft_utils import get_gptq_peft_model
+        from auto_gptq.utils.peft_utils import GPTQLoraModel
+        from ..plugin_utils.autogptq_utils import _create_new_module_triton, _replace_module
+
         peft_config, = modifiable_args # unpack modifiable args
 
         # some assertions
